requirements.txt
Flask==3.0.3
flask-cors==3.0.10
selenium>=4.0.0
webdriver-manager>=3.5.0
google-generativeai>=0.1.0
textblob